update file path for dummy data
