+++ src/utils/DummyData.tsx
@@ -1,0 +1,205 @@
+import * as React from 'react';
+import clsx from 'clsx';
+import { createStyles, Theme, withStyles, WithStyles, makeStyles } from '@material-ui/core/styles';
+import TableCell from '@material-ui/core/TableCell';
+import Paper from '@material-ui/core/Paper';
+import { AutoSizer, Column, Table, TableCellRenderer, TableHeaderProps } from 'react-virtualized';
+
+// Data
+import { SampleData } from './Data/SampleData'; 
+
+declare module '@material-ui/core/styles/withStyles' {
+  // Augment the BaseCSSProperties so that we can control jss-rtl
+  interface BaseCSSProperties {
+    /*
+     * Used to control if the rule-set should be affected by rtl transformation
+     */
+    flip?: boolean;
+  }
+}
+
+const styles = (theme: Theme) => 
+  createStyles({
+    head: {
+      backgroundColor: '#F7F8F9',
+      color: theme.palette.common.black,
+      width: '100%',
+    },
+    flexContainer: {
+      display: 'flex',
+      alignItems: 'center',
+      boxSizing: 'border-box',
+    },
+    table: {
+      // temporary right-to-left patch, waiting for
+      // https://github.com/bvaughn/react-virtualized/issues/454
+      '& .ReactVirtualized__Table__headerRow': {
+        flip: false,
+        paddingRight: theme.direction === 'rtl' ? '0 !important' : undefined,
+      },
+    },
+    tableRow: {
+      cursor: 'pointer',
+    },
+    tableRowHover: {
+      '&:hover': {
+        backgroundColor: theme.palette.grey[200],
+      },
+    },
+    tableCell: {
+      flex: 1,
+    },
+    noClick: {
+      cursor: 'initial',
+    },
+  });
+
+interface ColumnData {
+  dataKey: string;
+  label: string;
+  numeric?: boolean;
+  width: number;
+}
+
+interface Row {
+  index: number;
+}
+
+interface MuiVirtualizedTableProps extends WithStyles<typeof styles> {
+  columns: ColumnData[];
+  headerHeight?: number;
+  onRowClick?: () => void;
+  rowCount: number;
+  rowGetter: (row: Row) => Data;
+  rowHeight?: number;
+}
+
+class MuiVirtualizedTable extends React.PureComponent<MuiVirtualizedTableProps> {
+  static defaultProps = {
+    headerHeight: 48,
+    rowHeight: 48,
+  };
+
+  getRowClassName = ({ index }: Row) => {
+    const { classes, onRowClick } = this.props;
+
+    return clsx(classes.tableRow, classes.flexContainer, {
+      [classes.tableRowHover]: index !== -1 && onRowClick != null,
+    });
+  };
+
+  cellRenderer: TableCellRenderer = ({ cellData, columnIndex }: any) => {
+    const { columns, classes, rowHeight, onRowClick } = this.props;
+    return (
+      <TableCell
+        component="div"
+        className={clsx(classes.tableCell, classes.flexContainer, {
+          [classes.noClick]: onRowClick == null,
+        })}
+        variant="body"
+        style={{ height: rowHeight }}
+        align={(columnIndex != null && columns[columnIndex].numeric) || false ? 'right' : 'left'}
+      >
+        {cellData}
+      </TableCell>
+    );
+  };
+
+  headerRenderer = ({ label, columnIndex }: TableHeaderProps & { columnIndex: number }) => {
+    const { headerHeight, columns, classes } = this.props;
+
+    return (
+      <TableCell
+        component="div"
+        className={clsx(classes.head, classes.tableCell, classes.flexContainer, classes.noClick)}
+        variant="head"
+        style={{ height: headerHeight }}
+        align={columns[columnIndex].numeric || false ? 'right' : 'left'}
+      >
+        <span>{label}</span>
+      </TableCell>
+    );
+  };
+
+  render() {
+    const { classes, columns, rowHeight, headerHeight, ...tableProps } = this.props;
+    return (
+      <AutoSizer>
+        {({ height, width }: any) => (
+          <Table
+            height={height}
+            width={width}
+            rowHeight={rowHeight!}
+            gridStyle={{
+              direction: 'inherit',
+            }}
+            headerHeight={headerHeight!}
+            className={classes.table}
+            {...tableProps}
+            rowClassName={this.getRowClassName}
+          >
+            {columns.map(({ dataKey, ...other }, index) => {
+              return (
+                <Column
+                  key={dataKey}
+                  headerRenderer={(headerProps: any) =>
+                    this.headerRenderer({
+                      ...headerProps,
+                      columnIndex: index,
+                    })
+                  }
+                  className={classes.flexContainer}
+                  cellRenderer={this.cellRenderer}
+                  dataKey={dataKey}
+                  {...other}
+                />
+              );
+            })}
+          </Table>
+        )}
+      </AutoSizer>
+    );
+  }
+}
+
+export const VirtualizedTable = withStyles(styles)(MuiVirtualizedTable);
+
+// ---
+
+interface Data {
+  channels: string;
+  packetCommitment: string;
+  connections: string;
+  port: string;
+  id: number;
+  data: string;
+}
+type Sample = [string, string, string, string, string];
+
+const sample: Sample[] = [
+  ['connection name 1', 'channel 1', '6.0', '24', '4.0'],
+  ['connection name 2', 'channel 2', '9.0', '37', '4.3'],
+  ['connection name 3', 'channel 5', '16.0', '24', '6.0'],
+  ['connection name 4', 'channel 3', '3.7', '67', '4.3'],
+  ['connection name 5', 'channel 4', '16.0', '49', '3.9'],
+];
+
+// let sample: InstanceType<typeof SampleData>;
+
+function createData(
+  id: number,
+  connections: string,
+  channels: string,
+  port: string,
+  packetCommitment: string,
+  data: string,
+): Data {
+  return { id, connections, channels, port, packetCommitment, data };
+}
+
+export const rows: Data[] = [];
+
+for (let i = 0; i < 200; i += 1) {
+  const randomSelection = sample[Math.floor(Math.random() * sample.length)];
+  rows.push(createData(i, ...randomSelection));
+}
